Remove error raising implementation of getMode()
Due to the attributes 'modeFirefox', 'modeVLC' and 'modeImpress'
not existing (anymore) in the class 'paths', this function will
raise an AttributeError.

As the only function using getMode() is printMode() in
pyScreenCore.py, I decided to always return 'none'. We should
consider deleting this function entirely if it is not needed.